SJ-QuestionQualityClaw/pyproject.toml
[project]
name = "sjqqc"
version = "0.1.0"
description = "SJ-QuestionQualityClaw — Assessment question quality review agent"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # --- Core ---
    "pydantic>=2.5",
    "loguru>=0.7",
    "python-dotenv>=1.0",

    # --- LLM ---
    "openai>=1.10",
    "httpx>=0.27",

    # --- GitHub ---
    "PyGithub>=2.1",

    # --- Linear ---
    "gql[httpx]>=3.5",

    # --- Database ---
    "asyncpg>=0.29",
    "psycopg[binary]>=3.1",

    # --- Utilities ---
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.5",
    "mypy>=1.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring external services (GitHub, Linear, Postgres)",
    "slow: marks tests that take >10s",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
